requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
google-genai==1.50.1
python-dotenv==1.0.1
sse-starlette==2.1.0
python-multipart==0.0.9
pypdf==4.3.1
